# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniBid India API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform-wide admin stats
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [admin]
      summary: List all categories with stats (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Categories with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCategory"
  /admin/categories/{id}/floor:
    put:
      operationId: updateCategoryFloor
      tags: [admin]
      summary: Update a category price floor
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                floor:
                  type: number
              required: [floor]
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategory"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin, paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{id}:
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user trust/verification (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  isVerified:
                    type: boolean
                  trustScore:
                    type: number
                  omniScore:
                    type: number
                required: [id, name, isVerified, trustScore, omniScore]
components:
  schemas:
...
    AdminStats:
      type: object
      properties:
        users:
          type: object
          additionalProperties: true
        requirements:
          type: object
          additionalProperties: true
        bids:
          type: object
          additionalProperties: true
        payments:
          type: object
          additionalProperties: true
        disputes:
          type: object
          additionalProperties: true
        topSectors:
          type: array
          items:
            type: object
            additionalProperties: true
        topCities:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [users, requirements, bids, payments, disputes, topSectors, topCities]
    AdminCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        iconName:
          type: string
        priceFloor:
          type: number
        requirementCount:
          type: number
        bidCount:
          type: number
      required: [id, name, slug, iconName, priceFloor, requirementCount, bidCount]
    AdminUsersPage:
      type: object
      properties:
        users:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: number
        page:
          type: number
        pages:
          type: number
      required: [users, total, page, pages]
    AdminUserUpdate:
      type: object
      properties:
        isVerified:
          type: boolean
          nullable: true
        trustScore:
          type: number
          nullable: true
        omniScore:
          type: number
          nullable: true